# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCUMENT RWANDA",
        "version": "1.0"
    },
    "paths": {
        "/api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "description": "Liste des documents",
                "operationId": "listDocs",
                "responses": {
                    "200": {
                        "description": "Liste des documents."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "description": "Enregistrement d'un document",
                "operationId": "storeDocs",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "en",
                                    "fr",
                                    "description_en",
                                    "description_fr",
                                    "files"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "en": {
                                        "type": "string"
                                    },
                                    "fr": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "type": "string"
                                    },
                                    "description_en": {
                                        "type": "text"
                                    },
                                    "description_fr": {
                                        "type": "text"
                                    },
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "requestdoc_id": {
                                                    "type": "integer"
                                                },
                                                "required": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object",
                                            "example": "[1|1, 2|1, 3|0]"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document enregisté avec succès."
                    },
                    "400": {
                        "description": "Erreur de validation."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/documents/{uid}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "description": "Détail d'un document",
                "operationId": "showDocs",
                "responses": {
                    "200": {
                        "description": "Détail d'un document."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "description": "Modification d'un document",
                "operationId": "editDocs",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "en",
                                    "fr",
                                    "description_en",
                                    "description_fr",
                                    "files",
                                    "status"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "en": {
                                        "type": "string"
                                    },
                                    "fr": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "type": "string"
                                    },
                                    "description_en": {
                                        "type": "text"
                                    },
                                    "description_fr": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "requestdoc_id": {
                                                    "type": "integer"
                                                },
                                                "required": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object",
                                            "example": "[1|1, 2|1, 3|0]"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document modifié avec succès."
                    },
                    "400": {
                        "description": "Erreur de validation."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "description": "Suppression d'un document",
                "operationId": "deleteDocs",
                "responses": {
                    "200": {
                        "description": "Document supprimé avec succès."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}